+++ arancino/arancino_exceptions.py
@@ -48,3 +48,11 @@ class RedisGenericException(Exception):
         # Now for your custom code...
         self.error_code = error_code
 
+class NonCompatibilityException(Exception):
+    def __init__(self, message, error_code):
+
+        # Call the base class constructor with the parameters it needs
+        super(NonCompatibilityException, self).__init__(message)
+
+        # Now for your custom code...
+        self.error_code = error_code